Updated main and new functions
Main has been updated to include functions used to calculate the efficient frontier and Sharpe ratio. These are plotted in the main function to find the optimal portfolio. The optimal portfolio will be where the Sharpe ratio line touches. The next step for this is to find which portfolio weightings correspond to the portfolios in the graph, so we know which portfolio to choose.